// shared-lib/components/settings/components/account-and-personal-info/components/account-info/components/edit-password/models/edit-password.model.ts
export interface DisplayText {
  password: string;
  newPassword: string;
  confirmPassword: string;
  button: {
    save: string;
    cancel: string;
  };
  lastUpdate: string;
  alert: {
    message: string;
    imageUrl: string;
  };
  notificationMsg: string;
  manageNotification: string;
  actionOption: {
    headername: string;
    buttonLeft: string;
    buttonRight: string;
  };
}